Add look-up of `n` most common words in a text
I've added a method that allows you to look-up the `n` most common words
in a text. As `collections.Counter` breaks ties on occurrence rather
than the internal ordering of the keys, this method sorts the word
counts itself using a `key` function that ensures we first sort by
frequency (descending) and then by word (ascending, alphabetic).

This method raises a ValueError if a negative value for `n` was passed.