Add an implementation using modern Python idioms
I've added an alternative implementation of the WordFrequencyAnalyzer
using modern Python idioms instead of the Java-like API specified in the
assignment.

Most of the internal logic is the same, although the public API now
works with instances of the class and the word frequency analysis is
analyzed for any given analysis.